Remove sleep_us(1) in read_bit and write_bit
The 1 us sleep command does not work on pycom devices. It generates a wait time up to 30us. Without waiting time you get an impulse from at least 10us. Removing the waiting time improves significantly the stability of the communication of the new pycom devices as LoPy4, FiPy etc.
See whole case at https://forum.pycom.io/topic/2570/onewire-ds18x20-reading-not-stable-with-lopy4-no-problem-with-lopy/
Remark: There would be an additional improvement by adding gc.collect() to read_byte and write_byte, but to keep the lib as basic as possible I propose not to integrate it into the lib