onewire.py: Optimize timing, enable CRC check and slim the code
The reason for a change was the observation, that with about one in
300 measurements were wrong, some of them obvious. So I made a few
changes:

1. The timing of readbit was changed to shorten the initial low puls
to 2 µs by not calling sleep_us() at all and reading the state of the bit
about 8 µs later. That matches better the datasheet.

2. The CRC check was enabled in the method read_temp_async(). The function
returns None, when the CRC is wrong

3. The crc8() method is now table based, with two 16 byte tables.

4. Some inefficient optimizations in the onwire class were removed.
Especially caching system call names in local variables for calls
used only once in a functions was dropped.

With change 1, readings with wrong crc are very rare now, and enabling
the crc check allows to detect them.

The crc8 implementation is a Python port of the code published here:
http://lentz.com.au/blog/tag/crc-table-generator